Refactor build scripts
Avoid printing in the STDOUT/STDERR unless there is an error in the
command that was run, this would help improvie travis log file size,
as well as helping devs read the travis log.

Move commands to separate scripts instead of puting everything in the
the travis file, this helps future changes making it easier to
understand the build process.